# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EDA Explorer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload:
    post:
      operationId: uploadDataset
      tags: [datasets]
      summary: Upload a dataset file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "200":
          description: Dataset uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetInfo"
        "400":
          description: Invalid file
...
  /datasets/{datasetId}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get dataset info
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dataset info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetInfo"
        "404":
          description: Dataset not found
    delete:
      operationId: deleteDataset
      tags: [datasets]
      summary: Delete a dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /datasets/{datasetId}/preview:
    get:
      operationId: getDatasetPreview
      tags: [datasets]
      summary: Get first N rows of dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
        - name: rows
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Preview rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetPreview"
...
  /datasets/{datasetId}/transform:
    post:
      operationId: applyTransform
      tags: [datasets]
      summary: Apply transformation to dataset
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransformInput"
      responses:
        "200":
          description: Updated dataset info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetInfo"
  /datasets/{datasetId}/export/csv:
    get:
      operationId: exportCsv
      tags: [export]
      summary: Export dataset as CSV
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /datasets/{datasetId}/export/excel:
    get:
      operationId: exportExcel
      tags: [export]
      summary: Export summary stats as Excel
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /datasets/{datasetId}/export/outliers:
    get:
      operationId: exportOutliersCsv
      tags: [export]
      summary: Export outlier rows as CSV
      parameters:
        - name: datasetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV of outlier rows
          content:
            text/csv:
              schema:
                type: string
components:
  schemas:
...
    UploadInput:
      type: object
      properties:
        file:
          type: string
          format: binary
      required:
        - file
    DatasetInfo:
      type: object
      properties:
        id:
          type: string
        filename:
          type: string
        rows:
          type: integer
        cols:
          type: integer
        size_mb:
          type: number
        format:
          type: string
        encoding:
          type: string
        is_sampled:
          type: boolean
        sample_size:
          type: ["integer", "null"]
        created_at:
          type: string
      required:
        - id
        - filename
        - rows
        - cols
        - size_mb
        - format
        - encoding
        - is_sampled
        - created_at
    DatasetPreview:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
      required:
        - columns
        - rows
...
    TransformInput:
      type: object
      properties:
        column:
          type: string
        operation:
          type: string
          enum: [log, sqrt, one_hot, bin, drop_column, fill_mean, fill_median, fill_mode, fill_custom, drop_duplicates, change_dtype]
        params:
          type: object
          additionalProperties: true
      required:
        - column
        - operation